Refactoring `IEncryptionService`. Removing redundant code.
